최적화 관련 리팩토링 적용

+++ src/pages/signup/useSignup.ts
@@ -1,19 +1,12 @@
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
-import {
-  User,
-  createUserWithEmailAndPassword,
-  deleteUser,
-  updateProfile,
-} from 'firebase/auth';
-import { auth, db } from '@/firebase';
-import { FirestoreError, doc, setDoc } from 'firebase/firestore';
+import { createUserWithEmailAndPassword, deleteUser } from 'firebase/auth';
+import { auth } from '@/firebase/auth';
+import { FirestoreError } from 'firebase/firestore';
 import { FirebaseError } from 'firebase/app';
 import { MouseEventHandler } from 'react';
 import { toast } from 'sonner';
-import { UserSchema } from '@/types/FirebaseType';
 import { SignupFormDataType } from '@/types/FormType';
-import { getKoreanIsoDatetime } from '@/utils/utils';
 
 const useSignup = () => {
   const navigate = useNavigate();
@@ -29,53 +22,36 @@ const useSignup = () => {
   } = useForm<SignupFormDataType>();
 
   const submitLogic: SubmitHandler<SignupFormDataType> = async (data) => {
-    try {
-      const userCreateResult = await createUserWithEmailAndPassword(
-        auth,
-        data.email,
-        data.password,
-      );
-
-      const isoTime = getKoreanIsoDatetime();
-      const documentData: UserSchema = {
-        uid: userCreateResult.user.uid,
-        email: data.email,
-        accountType: data.accountType,
-        nickname: data.nickname,
-        createdAt: isoTime,
-        updatedAt: isoTime,
-      };
-
-      await setDoc(doc(db, 'user', documentData.uid), documentData);
-
-      await updateProfile(auth.currentUser as User, {
-        displayName: `${data.nickname}#${data.accountType}`,
-      });
-
-      toast.success('회원 가입 완료');
-    } catch (error: unknown) {
-      if (error instanceof FirebaseError) {
-        if (error.code === 'auth/email-already-in-use') {
-          toast.error('이미 가입된 이메일입니다.');
-        } else {
-          toast.error(error.message);
-        }
-      } else if (error instanceof FirestoreError) {
-        if (auth.currentUser) {
-          await deleteUser(auth.currentUser);
-          toast.error(
-            '가입 정보 등록에 실패했습니다. 다시 회원가입을 시도해주세요.',
-          );
-        }
-      } else {
-        toast.error((error as Error).message);
-      }
-    }
+    toast.promise(
+      createUserWithEmailAndPassword(auth, data.email, data.password),
+      {
+        loading: '회원가입입 요청을 처리중입니다...',
+        success: () => {
+          return '회원가입이 완료됐습니다.';
+        },
+        error: async (error) => {
+          console.error(error);
+          if (error instanceof FirebaseError) {
+            if (error.code === 'auth/email-already-in-use') {
+              return '이미 가입된 이메일입니다.';
+            } else {
+              if (error instanceof FirebaseError && auth.currentUser) {
+                await deleteUser(auth.currentUser);
+              }
+              return `회원가입에 실패했습니다. 다시 시도해주세요. [${error.message}]`;
+            }
+          } else if (error instanceof FirestoreError) {
+            if (auth.currentUser) {
+              await deleteUser(auth.currentUser);
+              return '가입 정보 등록에 실패했습니다. 다시 회원가입을 시도해주세요.';
+            }
+          } else {
+            return (error as Error).message;
+          }
+        },
+      },
+    );
   };
-
-  const registerAccountType = register('accountType', {
-    required: '계정 종류 선택은 필수입니다.',
-  });
 
   const registerEmail = register('email', {
     required: '아이디는 필수 입력입니다.',
@@ -98,32 +74,21 @@ const useSignup = () => {
     required: '패스워드는 필수 입력입니다.',
   });
 
-  const registerNickname = register('nickname', {
-    minLength: {
-      value: 2,
-      message: '닉네임은 2자 이상이어야 합니다.',
-    },
-    maxLength: {
-      value: 10,
-      message: '닉네임은 10자 이내여야 합니다.',
-    },
-    pattern: {
-      value: /^[0-9a-zA-Z가-힣\x20]*$/,
-      message: '한글/영어/숫자만 사용해주세요.',
-    },
-    required: '닉네임은 필수 입력입니다.',
-  });
-
-  const accountTypeOptions = [
-    {
-      value: '구매자',
-      label: '구매자 계정',
-    },
-    {
-      value: '판매자',
-      label: '판매자 계정',
-    },
-  ];
+  // const registerNickname = register('nickname', {
+  //   minLength: {
+  //     value: 2,
+  //     message: '닉네임은 2자 이상이어야 합니다.',
+  //   },
+  //   maxLength: {
+  //     value: 10,
+  //     message: '닉네임은 10자 이내여야 합니다.',
+  //   },
+  //   pattern: {
+  //     value: /^[0-9a-zA-Z가-힣\x20]*$/,
+  //     message: '한글/영어/숫자만 사용해주세요.',
+  //   },
+  //   required: '닉네임은 필수 입력입니다.',
+  // });
 
   return {
     redirectToSignin,
@@ -131,11 +96,8 @@ const useSignup = () => {
     submitLogic,
     isSubmitting,
     errors,
-    registerAccountType,
     registerEmail,
     registerPassword,
-    registerNickname,
-    accountTypeOptions,
   };
 };
